+++ packages/client/Atmosphere.ts
@@ -363,6 +363,7 @@ export default class Atmosphere extends Environment {
     if (!this.authObj) return
     const {exp, sub: viewerId, rol, iat} = this.authObj
     if (rol === 'impersonate') {
+      this.viewerId = viewerId
       return this.validateImpersonation(iat)
     }
     // impersonation token must be < 10 seconds old (ie log them out automatically)