+++ packages/client/components/SnackbarMessage.tsx
@@ -4,7 +4,9 @@ import {TransitionStatus} from '../hooks/useTransition'
 import {DECELERATE} from '../styles/animation'
 import {snackbarShadow} from '../styles/elevation'
 import {PALETTE} from '../styles/paletteV3'
+import {ICON_SIZE} from '../styles/typographyV2'
 import {Radius, ZIndex} from '../types/constEnums'
+import Icon from './Icon'
 import {SnackAction} from './Snackbar'
 import SnackbarMessageAction from './SnackbarMessageAction'
 
@@ -17,6 +19,7 @@ interface Props {
   secondaryAction?: SnackAction
   onMouseEnter: () => void
   onMouseLeave: () => void
+  showDismissButton?: boolean
 }
 
 const Space = styled('div')({
@@ -38,13 +41,32 @@ const MessageStyles = styled('div')<{status: TransitionStatus}>(({status}) => ({
   padding: 8,
   transition: `all 300ms ${DECELERATE}`,
   opacity: status === TransitionStatus.MOUNTED || status === TransitionStatus.EXITING ? 0 : 1,
-  transform: `translateY(${status === TransitionStatus.MOUNTED ? 20 : status === TransitionStatus.EXITING ? -20 : 0
-    }px)`,
+  transform: `translateY(${
+    status === TransitionStatus.MOUNTED ? 20 : status === TransitionStatus.EXITING ? -20 : 0
+  }px)`,
   pointerEvents: 'auto',
   userSelect: 'none',
   zIndex: ZIndex.SNACKBAR
 }))
 
+const DismissButton = styled('button')({
+  border: 'none',
+  backgroundColor: 'inherit',
+  marginLeft: '8px',
+  cursor: 'pointer',
+  padding: 5,
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center'
+})
+
+const StyledIcon = styled(Icon)({
+  color: PALETTE.SLATE_500,
+  fontSize: ICON_SIZE.MD18,
+  '&:hover': {
+    opacity: 0.5
+  }
+})
 const useTransitionEnd = (
   timeout: number,
   status: TransitionStatus,
@@ -64,20 +86,26 @@ const SnackbarMessage = (props: Props) => {
     onTransitionEnd,
     dismissSnack,
     onMouseEnter,
-    onMouseLeave
+    onMouseLeave,
+    showDismissButton
   } = props
   useTransitionEnd(300, status, onTransitionEnd)
   return (
     <Space>
       <MessageStyles
         status={status}
-        onClick={dismissSnack}
         onMouseEnter={onMouseEnter}
         onMouseLeave={onMouseLeave}
+        onClick={dismissSnack}
       >
         <Text>{message}</Text>
         <SnackbarMessageAction action={action} />
         <SnackbarMessageAction action={secondaryAction} />
+        {showDismissButton && (
+          <DismissButton onClick={dismissSnack}>
+            <StyledIcon>close</StyledIcon>
+          </DismissButton>
+        )}
       </MessageStyles>
     </Space>
   )